Refactor connection status check and query execution
This commit refactors the code in the `LoaderConnectionStatusCheck` and `RunLoaderQuery` classes. It introduces a new dependency on the `Process` facade from the `Illuminate\Support\Facades` namespace. The code now checks for the presence of a VPN password in the application configuration and runs a command using the `snx` utility if the password is available. After executing the query, the code also runs another command to disconnect from the VPN if a password was provided.

The `config/app.php` file has also been modified to include a new configuration option `vpn_password` which can be set in the environment file.

The `routes/web.php` file has been updated to remove duplicate resource routes for `loader-jobs`.